% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRL_preprocess.R
\name{PRL_preprocess_csv}
\alias{PRL_preprocess_csv}
\title{preprocess for CSV file of PRL task}
\usage{
PRL_preprocess_csv(name_of_csv_file)
}
\arguments{
\item{name_of_csv_file}{name of files of data}
}
\value{
preprocessed data
trial: 1,2,3...
choice: 1=select S1, 2=select S2
rt: response time
reward: reward
}
\description{
preprocess for CSV file of PRL task
}
\examples{
# PRL_preprocess_csv("data.csv")
}
